+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/dividab/graphql-norm-patch/compare/v0.17.0...master)
 
+### Changed
+
+- Upgrade to 0.4.0 of graphql-norm-stale.
+
 ## [v0.17.0](https://github.com/dividab/graphql-norm-patch/compare/v0.16.0...v0.17.0) - 2019-10-02
 
 ### Added